added screenshot capture

+++ src/main/java/stepDefinitions/MasterHooks.java
@@ -1,5 +1,10 @@
 package stepDefinitions;
 
+import org.openqa.selenium.TakesScreenshot;
+
+import com.mongodb.MapReduceCommand.OutputType;
+
+import cucumber.api.Scenario;
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import utils.DriverFactory;
@@ -12,11 +17,22 @@ public class MasterHooks extends DriverFactory {
 	}
 	
 	@After
-	public void tearDown() {
-		if(driver != null) {
-			driver.manage().deleteAllCookies();
-			driver.quit();
-			driver = null;
+	public void tearDownAndScreenshotOnFailure(Scenario scenario) {
+		try {
+			if (driver != null && scenario.isFailed()) {
+				scenario.embed(((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES), "image/png");
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver = null;
+			}
+
+			if (driver != null) {
+				driver.manage().deleteAllCookies();
+				driver.quit();
+				driver = null;
+			}
+		} catch (Exception e) {
+			System.out.println("Method Failed: tearDownAndScreenshotOnFailure, Exception: " + e.getMessage());
 		}
 	}
